requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
cffi==1.15.0
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.9
django-debug-toolbar==3.2.2
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.4
parsel==1.6.0
Pillow==8.4.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
queuelib==1.6.2
Scrapy==2.5.1
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyrt==0.12.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
w3lib==1.22.0
whitenoise==5.3.0
zope.interface==5.4.0
